requirements.txt
colorama
requests
sentencepiece
tqdm
pybars3
pyyaml
sseclient-py
typing_extensions
sentence_transformers